requirements.txt
python-telegram-bot==20.5
apscheduler==3.10.4
asyncio==3.4.3
requests==2.31.0